moved and replaced files from voters app to users app
