+++ users/apps.py
@@ -1,6 +1,9 @@
 from django.apps import AppConfig
 
 
-class UsersConfig(AppConfig):
+class VotersConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'users'
+    name = 'voters'
+
+    def ready(self):
+        import voters.signals
