--- voters/apps.py
@@ -1,9 +1,0 @@
-from django.apps import AppConfig
-
-
-class VotersConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'voters'
-
-    def ready(self):
-        import voters.signals
