--- voters/templates/voters/login.html
@@ -1,51 +1,0 @@
-{% extends 'voters/base.html' %}{% block title %}Login{% endblock %}
-{% load crispy_forms_tags %}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="utf-8">
-	<meta content="width=device-width, initial-scale=1.0" name="viewport">
-	<meta content="" name="description">
-	<meta content="" name="keywords">
-</head>
-<body>
-	{% block content %}
-	<main>
-		<div class="container">
-			<section class="section register min-vh-100 d-flex flex-column align-items-center justify-content-center py-4">
-				<div class="container">
-					<div class="row justify-content-center">
-						<div class="col-lg-6">
-							<div class="card mb-3">
-								<div class="card-body">
-									<div class="pt-4 pb-2">
-										<h5 class="card-title text-center pb-0 fs-4">Login to Your Account</h5>
-									</div>
-									<form action="" method="post" class="p-3">
-										{% csrf_token %}
-										<div class="row">
-											<div class="col-12">
-												{% include 'messages.html' %}{{form|crispy}}
-											</div>
-										</div>
-										<div class="col-12 mt-3">
-											<button class="btn btn-primary w-100" type="submit">Login</button>
-										</div>
-										<div class="col-12">
-											<hr class="dropdown-divider">
-											<h6 class="text-center"><a href="">Forgot password?</a></h6>
-											<p class="small mb-0 text-center">Don't have account? <a href="{% url 'voters_signup' %}">Create an account</a></p>
-										</div>
-									</form>
-								</div>
-							</div>
-
-						</div>
-					</div>
-				</div>
-			</section>
-		</div>
-	</main>
-  	{% endblock %}
-</body>
-</html>
