--- voters/templates/voters/profile.html
@@ -1,155 +1,0 @@
-{% extends 'voters/base.html' %}{% block title %}Profile{% endblock %}
-{% load crispy_forms_tags %}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="utf-8">
-	<meta content="width=device-width, initial-scale=1.0" name="viewport">
-</head>
-<body>
-	{% block content %}
-	<main id="main" class="main">
-		<section class="section profile">
-			<div class="row">
-				<div class="col-xl-4">
-					<div class="card">
-						<div class="card-body profile-card pt-4 d-flex flex-column align-items-center">
-							<img src="{{request.user.voters.profile_pic.url}}" alt="Profile" class="rounded-circle">
-							<h2>{{request.user.voters.voter|title}} {% if request.user.voters.registered is True %}<i class="bi-check-circle-fill" style="color:#13b46c"></i>{% endif %}</h2>
-							<h3>{{request.user.voters.year}}</h3>
-						</div>
-					</div>
-
-				</div>
-
-				<div class="col-xl-8">
-					<div class="card">
-						<div class="card-body pt-3">
-							<ul class="nav nav-tabs nav-tabs-bordered">
-
-								<li class="nav-item">
-									<button class="nav-link active" data-bs-toggle="tab" data-bs-target="#profile-overview">Overview</button>
-								</li>
-
-								<li class="nav-item">
-									<button class="nav-link" data-bs-toggle="tab" data-bs-target="#profile-edit">{% if request.user.voters.registered is False %}Get Registered <i class="bi-exclamation-circle-fill" style="color: #e71b1b;"></i>{% else %}Edit Profile{% endif %}</button>
-								</li>
-								{% if request.user.voters.registered is True %}
-								<li class="nav-item">
-									<button class="nav-link" data-bs-toggle="tab" data-bs-target="#profile-change-password">Change Password</button>
-								</li>
-								{% endif %}
-
-							</ul>
-							<div class="tab-content pt-2">
-								{% include 'messages.html' %}
-								<div class="tab-pane fade show active profile-overview" id="profile-overview">
-									<h5 class="card-title">Profile Details</h5>
-
-									<div class="row">
-										<div class="col-lg-3 col-md-4 label">Full Name</div>
-										<div class="col-lg-9 col-md-8">{{request.user.voters.voter|title}}</div>
-									</div>
-
-									<div class="row">
-										<div class="col-lg-3 col-md-4 label">Gender</div>
-										<div class="col-lg-9 col-md-8">{{request.user.voters.gender|default:" -- "}}</div>
-									</div>
-
-									<div class="row">
-										<div class="col-lg-3 col-md-4 label">Age</div>
-										<div class="col-lg-9 col-md-8">{{request.user.voters.age|default:" -- "}}</div>
-									</div>
-
-									<div class="row">
-										<div class="col-lg-3 col-md-4 label">Date of birth</div>
-										<div class="col-lg-9 col-md-8">{{request.user.voters.dob|default:" -- "}}</div>
-									</div>
-									
-									<div class="row">
-										<div class="col-lg-3 col-md-4 label">Phone No.</div>
-										<div class="col-lg-9 col-md-8">{{request.user.voters.phone_no|default:" -- "}}</div>
-									</div>
-
-									<div class="row">
-										<div class="col-lg-3 col-md-4 label">School</div>
-										<div class="col-lg-9 col-md-8">{{request.user.voters.school|default:" -- "}}</div>
-									</div>
-
-									<div class="row">
-										<div class="col-lg-3 col-md-4 label">Year of study</div>
-										<div class="col-lg-9 col-md-8">{{request.user.voters.year|default:" -- "}}</div>
-									</div>
-
-									<div class="row">
-										<div class="col-lg-3 col-md-4 label">Semester</div>
-										<div class="col-lg-9 col-md-8">{{request.user.voters.semester|default:" -- "}}</div>
-									</div>
-
-								</div>
-
-								<div class="tab-pane fade profile-edit pt-3" id="profile-edit">
-									<form action="" method="post" enctype="multipart/form-data">
-										{% csrf_token %}
-										<div class="row mb-3">
-											<label for="image" class="col-md-4 col-lg-3 col-form-label">Profile Image</label>
-											<div class="col-md-8 col-lg-9">
-												<img src="{{request.user.voters.profile_pic.url}}" alt="Profile">
-											</div>
-										</div>
-										<div class="row mb-2">
-											<div class="col-md-8 col-lg-12">
-												{% if request.user.voters.gender == "" %}
-												{{UpdateProfileForm|crispy}}
-												{% else %}
-												{{EditProfileForm|crispy}}
-												{% endif %}
-											</div>
-										</div>
-										<div class="text-center d-grid gap-2">
-											<button type="submit" class="btn btn-primary">Save Changes</button>
-										</div>
-									</form>
-
-								</div>
-
-								<div class="tab-pane fade pt-3" id="profile-change-password">
-									<form action="" method="post">
-
-										<div class="row mb-3">
-										<label for="currentPassword" class="col-md-4 col-lg-3 col-form-label">Current Password</label>
-										<div class="col-md-8 col-lg-9">
-											<input name="password" type="password" class="form-control" id="currentPassword">
-										</div>
-										</div>
-
-										<div class="row mb-3">
-										<label for="newPassword" class="col-md-4 col-lg-3 col-form-label">New Password</label>
-										<div class="col-md-8 col-lg-9">
-											<input name="newpassword" type="password" class="form-control" id="newPassword">
-										</div>
-										</div>
-
-										<div class="row mb-3">
-										<label for="renewPassword" class="col-md-4 col-lg-3 col-form-label">Re-enter New Password</label>
-										<div class="col-md-8 col-lg-9">
-											<input name="renewpassword" type="password" class="form-control" id="renewPassword">
-										</div>
-										</div>
-
-										<div class="text-center">
-										<button type="submit" class="btn btn-primary">Change Password</button>
-										</div>
-									</form>
-
-								</div>
-							</div>
-						</div>
-					</div>
-				</div>
-			</div>
-		</section>
-	</main>
-	{% endblock %}
-</body>
-</html>
