--- voters/templates/voters/signup.html
@@ -1,50 +1,0 @@
-{% extends 'voters/base.html' %}{% block title %}Login{% endblock %}
-{% load crispy_forms_tags %}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="utf-8">
-	<meta content="width=device-width, initial-scale=1.0" name="viewport">
-	<meta content="" name="description">
-	<meta content="" name="keywords">
-</head>
-<body>
-	{% block content %}
-	<main>
-		<div class="container signup">
-			<section class="section register min-vh-100 d-flex flex-column align-items-center justify-content-center py-4">
-				<div class="container">
-					<div class="row justify-content-center">
-						<div class="col-lg-6">
-							<div class="card mb-3">
-								<div class="card-body">
-									<div class="pt-4 pb-2">
-										<h5 class="card-title text-center pb-0 fs-4">Create Account</h5>
-									</div>
-									<form action="" method="post" class="p-3">
-										{% csrf_token %}
-										<div class="row">
-											<div class="col-12">
-												{% include 'messages.html' %}{{signup_form|crispy}}
-											</div>
-										</div>
-										<div class="col-12 mt-3">
-											<button class="btn btn-success w-100" type="submit">Sign Up</button>
-										</div>
-										<div class="col-12">
-											<hr class="dropdown-divider">
-											<p class="small mb-0 text-center">Already have account? Click here to <a href="{% url 'voters_login' %}">login</a></p>
-										</div>
-									</form>
-								</div>
-							</div>
-
-						</div>
-					</div>
-				</div>
-			</section>
-		</div>
-	</main>
-  	{% endblock %}
-</body>
-</html>
